Rewrote math logic + added previous calculation
placed all the main functions at the end and then
redid all the events when a button is clicked
since the logic behind them was too long and
repetitive. Added a total display to the screen
and not just one number or operand at a time.
Additionally, used the past calculation to display
it above current calculation