Add backspace and sign change buttons + change UI
create a button for backspace and sign change
so the user could perform actions that were only
exclusive to the keyboard at the time. Also,
changed the UI a bit so that the stickers are
on a different level of the calculator input to
not mess with the symettrical 5x4 button organiza-
tion.